Update spec for how this failure now returns from the Contracts gem
I can't pin down exactly where in the contracts.ruby changelog this happened but I believe
this is the correct patch for this version of contracts.ruby. I bopped around in the debugger
for a while and this looks to be the correct test tweak. This won't backport to ruby2 without
a conditional.