feat(gems): add non-default gems to the gemspec
Starting in Ruby 3.5.0, the `logger` and `ostruct` will not be part of
the default gem and will need to be manually vendored.

Example output with warnings:

```
/entitlements/vendor/gems/ruby/3.3.0/gems/concurrent-ruby-1.3.3/lib/concurrent-ruby/concurrent/concern/deprecation.rb:1: warning: logger was loaded from the standard library, but will no longer be part of the default gems starting from Ruby 3.5.0.
You can add logger to your Gemfile or gemspec to silence this warning.
/entitlements/bin/deploy-entitlements:27: warning: ostruct was loaded from the standard library, but will no longer be part of the default gems starting from Ruby 3.5.0.
You can add ostruct to your Gemfile or gemspec to silence this warning.
```